+++ src/main/java/io/spring/api/article/ArticleApi.java
@@ -1,11 +1,11 @@
-package io.spring.api;
+package io.spring.api.article;
 
 import io.spring.api.exception.NoAuthorizationException;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.ArticleQueryService;
 import io.spring.application.article.ArticleCommandService;
-import io.spring.application.article.UpdateArticleParam;
-import io.spring.application.data.ArticleData;
+import io.spring.api.article.request.UpdateArticleParam;
+import io.spring.api.article.response.ArticleData;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.service.AuthorizationService;
