+++ src/main/java/io/spring/api/article/ArticleFavoriteApi.java
@@ -1,8 +1,8 @@
-package io.spring.api;
+package io.spring.api.article;
 
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.ArticleQueryService;
-import io.spring.application.data.ArticleData;
+import io.spring.api.article.response.ArticleData;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.favorite.ArticleFavorite;
