+++ src/main/java/io/spring/api/article/request/UpdateArticleParam.java
@@ -1,4 +1,4 @@
-package io.spring.application.article;
+package io.spring.api.article.request;
 
 import com.fasterxml.jackson.annotation.JsonRootName;
