+++ src/main/java/io/spring/api/article/response/ArticleData.java
@@ -1,9 +1,11 @@
-package io.spring.application.data;
+package io.spring.api.article.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.spring.application.DateTimeCursor;
 import java.time.Instant;
 import java.util.List;
+
+import io.spring.api.user.response.ProfileData;
 import lombok.*;
 
 @Getter
