+++ src/main/java/io/spring/api/article/response/ArticleFavoriteCount.java
@@ -1,3 +1,3 @@
-package io.spring.application.data;
+package io.spring.api.article.response;
 
 public record ArticleFavoriteCount(String id, Integer count) {}
