+++ src/main/java/io/spring/api/comment/CommentsApi.java
@@ -1,12 +1,10 @@
 package io.spring.api.comment;
 
-import com.fasterxml.jackson.annotation.JsonRootName;
-
 import io.spring.api.comment.request.NewCommentParam;
 import io.spring.api.exception.NoAuthorizationException;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.CommentQueryService;
-import io.spring.application.data.CommentData;
+import io.spring.api.comment.response.CommentData;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.comment.Comment;
@@ -17,7 +15,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
+
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
