+++ src/main/java/io/spring/api/user/request/UpdateUserCommand.java
@@ -1,5 +1,6 @@
-package io.spring.application.user;
+package io.spring.api.user.request;
 
+import io.spring.application.user.UpdateUserConstraint;
 import io.spring.core.user.User;
 
 @UpdateUserConstraint
