+++ src/main/java/io/spring/api/user/response/UserWithToken.java
@@ -1,4 +1,4 @@
-package io.spring.application.data;
+package io.spring.api.user.response;
 
 public record UserWithToken(String email, String username, String bio, String image, String token) {
   public static UserWithToken of(UserData userData, String token) {
