+++ src/main/java/io/spring/application/ArticleQueryService.java
@@ -2,9 +2,9 @@ package io.spring.application;
 
 import static java.util.stream.Collectors.toList;
 
-import io.spring.application.data.ArticleData;
-import io.spring.application.data.ArticleDataList;
-import io.spring.application.data.ArticleFavoriteCount;
+import io.spring.api.article.response.ArticleData;
+import io.spring.api.article.response.ArticleDataList;
+import io.spring.api.article.response.ArticleFavoriteCount;
 import io.spring.core.user.User;
 import io.spring.infrastructure.mybatis.readservice.ArticleFavoritesReadService;
 import io.spring.infrastructure.mybatis.readservice.ArticleReadService;
