+++ src/main/java/io/spring/application/CommentQueryService.java
@@ -1,6 +1,6 @@
 package io.spring.application;
 
-import io.spring.application.data.CommentData;
+import io.spring.api.comment.response.CommentData;
 import io.spring.core.user.User;
 import io.spring.infrastructure.mybatis.readservice.CommentReadService;
 import io.spring.infrastructure.mybatis.readservice.UserRelationshipQueryService;
