+++ src/main/java/io/spring/graphql/ArticleMutation.java
@@ -7,8 +7,8 @@ import graphql.execution.DataFetcherResult;
 import io.spring.api.exception.NoAuthorizationException;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.article.ArticleCommandService;
-import io.spring.application.article.NewArticleParam;
-import io.spring.application.article.UpdateArticleParam;
+import io.spring.api.article.request.NewArticleParam;
+import io.spring.api.article.request.UpdateArticleParam;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.favorite.ArticleFavorite;
