+++ src/main/java/io/spring/graphql/CommentDatafetcher.java
@@ -12,8 +12,8 @@ import io.spring.application.CursorPageParameter;
 import io.spring.application.CursorPager;
 import io.spring.application.CursorPager.Direction;
 import io.spring.application.DateTimeCursor;
-import io.spring.application.data.ArticleData;
-import io.spring.application.data.CommentData;
+import io.spring.api.article.response.ArticleData;
+import io.spring.api.comment.response.CommentData;
 import io.spring.core.user.User;
 import io.spring.graphql.DgsConstants.ARTICLE;
 import io.spring.graphql.DgsConstants.COMMENTPAYLOAD;
