+++ src/main/java/io/spring/graphql/CommentMutation.java
@@ -7,7 +7,7 @@ import graphql.execution.DataFetcherResult;
 import io.spring.api.exception.NoAuthorizationException;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.CommentQueryService;
-import io.spring.application.data.CommentData;
+import io.spring.api.comment.response.CommentData;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.comment.Comment;
