+++ src/main/java/io/spring/graphql/MeDatafetcher.java
@@ -6,8 +6,8 @@ import graphql.execution.DataFetcherResult;
 import graphql.schema.DataFetchingEnvironment;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.UserQueryService;
-import io.spring.application.data.UserData;
-import io.spring.application.data.UserWithToken;
+import io.spring.api.user.response.UserData;
+import io.spring.api.user.response.UserWithToken;
 import io.spring.core.service.JwtService;
 import io.spring.graphql.DgsConstants.QUERY;
 import io.spring.graphql.DgsConstants.USERPAYLOAD;
