+++ src/main/java/io/spring/graphql/ProfileDatafetcher.java
@@ -6,9 +6,9 @@ import com.netflix.graphql.dgs.InputArgument;
 import graphql.schema.DataFetchingEnvironment;
 import io.spring.api.exception.ResourceNotFoundException;
 import io.spring.application.ProfileQueryService;
-import io.spring.application.data.ArticleData;
-import io.spring.application.data.CommentData;
-import io.spring.application.data.ProfileData;
+import io.spring.api.article.response.ArticleData;
+import io.spring.api.comment.response.CommentData;
+import io.spring.api.user.response.ProfileData;
 import io.spring.core.user.User;
 import io.spring.graphql.DgsConstants.ARTICLE;
 import io.spring.graphql.DgsConstants.COMMENT;
