+++ src/main/java/io/spring/infrastructure/mybatis/readservice/ArticleReadService.java
@@ -2,7 +2,7 @@ package io.spring.infrastructure.mybatis.readservice;
 
 import io.spring.application.CursorPageParameter;
 import io.spring.application.Page;
-import io.spring.application.data.ArticleData;
+import io.spring.api.article.response.ArticleData;
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
