+++ src/main/java/io/spring/infrastructure/mybatis/readservice/UserReadService.java
@@ -1,6 +1,6 @@
 package io.spring.infrastructure.mybatis.readservice;
 
-import io.spring.application.data.UserData;
+import io.spring.api.user.response.UserData;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
