+++ src/main/resources/mapper/TransferData.xml
@@ -1,21 +1,21 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 <mapper namespace="transfer.data">
-    <resultMap id="favoriteCount" type="io.spring.application.data.ArticleFavoriteCount">
+    <resultMap id="favoriteCount" type="io.spring.api.article.response.ArticleFavoriteCount">
         <constructor>
             <idArg column="id" javaType="string"/>
             <arg column="favoriteCount" javaType="integer"/>
         </constructor>
     </resultMap>
 
-    <resultMap id="profileData" type="io.spring.application.data.ProfileData">
+    <resultMap id="profileData" type="io.spring.api.user.response.ProfileData">
         <id column="userId" property="id"/>
         <result column="userUsername" property="username"/>
         <result column="userBio" property="bio"/>
         <result column="userImage" property="image"/>
     </resultMap>
 
-    <resultMap id="articleData" type="io.spring.application.data.ArticleData">
+    <resultMap id="articleData" type="io.spring.api.article.response.ArticleData">
         <id column="articleId" property="id"/>
         <result column="articleSlug" property="slug"/>
         <result column="articleTitle" property="title"/>
@@ -29,7 +29,7 @@
         </collection>
     </resultMap>
 
-    <resultMap id="commentData" type="io.spring.application.data.CommentData">
+    <resultMap id="commentData" type="io.spring.api.comment.response.CommentData">
         <id column="commentId" property="id"/>
         <result column="commentBody" property="body"/>
         <result column="commentCreatedAt" property="createdAt"/>
