+++ src/main/resources/mapper/UserReadService.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
 <mapper namespace="io.spring.infrastructure.mybatis.readservice.UserReadService">
-    <select id="findByUsername" resultType="io.spring.application.data.UserData">
+    <select id="findByUsername" resultType="io.spring.api.user.response.UserData">
         select * from users where username = #{username}
     </select>
-    <select id="findById" resultType="io.spring.application.data.UserData">
+    <select id="findById" resultType="io.spring.api.user.response.UserData">
         select * from users where id = #{id}
     </select>
 </mapper>
