+++ src/test/java/io/spring/api/ArticlesApiTest.java
@@ -10,11 +10,12 @@ import static org.mockito.Mockito.when;
 
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import io.spring.JacksonCustomizations;
+import io.spring.api.article.ArticlesApi;
 import io.spring.api.security.WebSecurityConfig;
 import io.spring.application.ArticleQueryService;
 import io.spring.application.article.ArticleCommandService;
-import io.spring.application.data.ArticleData;
-import io.spring.application.data.ProfileData;
+import io.spring.api.article.response.ArticleData;
+import io.spring.api.user.response.ProfileData;
 import io.spring.core.article.Article;
 import java.time.Instant;
 import java.util.HashMap;
