+++ src/test/java/io/spring/api/ProfileApiTest.java
@@ -9,8 +9,9 @@ import static org.mockito.Mockito.when;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import io.spring.JacksonCustomizations;
 import io.spring.api.security.WebSecurityConfig;
+import io.spring.api.user.ProfileApi;
 import io.spring.application.ProfileQueryService;
-import io.spring.application.data.ProfileData;
+import io.spring.api.user.response.ProfileData;
 import io.spring.core.user.FollowRelation;
 import io.spring.core.user.User;
 import java.util.Optional;
