+++ src/test/java/io/spring/api/TestWithCurrentUser.java
@@ -3,7 +3,7 @@ package io.spring.api;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 
-import io.spring.application.data.UserData;
+import io.spring.api.user.response.UserData;
 import io.spring.core.service.JwtService;
 import io.spring.core.user.User;
 import io.spring.core.user.UserRepository;
