+++ src/test/java/io/spring/api/UsersApiTest.java
@@ -10,8 +10,9 @@ import static org.mockito.Mockito.when;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import io.spring.JacksonCustomizations;
 import io.spring.api.security.WebSecurityConfig;
+import io.spring.api.user.UsersApi;
 import io.spring.application.UserQueryService;
-import io.spring.application.data.UserData;
+import io.spring.api.user.response.UserData;
 import io.spring.application.user.UserService;
 import io.spring.core.service.JwtService;
 import io.spring.core.user.User;
