+++ src/test/java/io/spring/application/article/ArticleQueryServiceTest.java
@@ -6,8 +6,8 @@ import io.spring.application.CursorPager;
 import io.spring.application.CursorPager.Direction;
 import io.spring.application.DateTimeCursor;
 import io.spring.application.Page;
-import io.spring.application.data.ArticleData;
-import io.spring.application.data.ArticleDataList;
+import io.spring.api.article.response.ArticleData;
+import io.spring.api.article.response.ArticleDataList;
 import io.spring.core.article.Article;
 import io.spring.core.article.ArticleRepository;
 import io.spring.core.favorite.ArticleFavorite;
