+++ src/test/java/io/spring/application/profile/ProfileQueryServiceTest.java
@@ -1,7 +1,7 @@
 package io.spring.application.profile;
 
 import io.spring.application.ProfileQueryService;
-import io.spring.application.data.ProfileData;
+import io.spring.api.user.response.ProfileData;
 import io.spring.core.user.User;
 import io.spring.core.user.UserRepository;
 import io.spring.infrastructure.DbTestBase;
